+++ Rdmp.Core/DataExport/Data/Project.cs
@@ -176,7 +176,14 @@ public class Project : DatabaseEntity, IProject, ICustomSearchString, ICheckable
     {
         var associations =
             Repository.GetAllObjectsWithParent<ProjectCohortIdentificationConfigurationAssociation>(this);
-        return associations.Select(a => a.CohortIdentificationConfiguration).Where(c => c != null).ToArray();
+        return associations.Select(a => a.CohortIdentificationConfiguration).Where(c => c != null && !c.IsTemplate).ToArray();
+    }
+
+    public CohortIdentificationConfiguration[] GetAssociatedTemplateCohortIdentificationConfigurations()
+    {
+        var associations =
+            Repository.GetAllObjectsWithParent<ProjectCohortIdentificationConfigurationAssociation>(this);
+        return associations.Select(a => a.CohortIdentificationConfiguration).Where(c => c != null && c.IsTemplate).ToArray();
     }
 
     /// <summary>