+++ CHANGELOG.md
@@ -4,6 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [9.1.0] - 2025-11-24
+- Fix bug with duplicate searchables
+- Improve UI for committing cohorts across projects
+- Add the ability to Template Cohort Identification Configurations (see [Documentation\cohorts\CohortIdentificationConfigurationTemplates.md])
+- Require all deletes to enter a commit message when using the commit system
+- Introduce ability to view Catalogues in a flat view
+- Fix bug where Internal catalogues were still able to be added to a CIC
+- Add ability to use archive trigger when re-releasing to a database
+
 ## [9.0.3] - 2025-11-03
 - Improve checking for default pipelines
 - Improve confluence catalogue metadata extraction
@@ -11,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [9.0.2] - 2025-10-21
 - Add ability to view logs as a flat list. Default settings can be updated via user settings
+- allow the use of Catalogue CIC filters in corresponding Extractions
 - Add Data Load component to allow SQL to be executed on an external database server
 - Improve Extraction Log Viewer Filter
 - Add Dataset Variable document to extractions